backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
aiosqlite
python-dotenv
websockets
tiktoken
pyotp
python-jose[cryptography]
passlib
bcrypt==4.0.1